+++ src/ansibleguy-webui/aw/api_endpoints/credentials.py
@@ -9,7 +9,7 @@ from aw.model.job import Job, JobExecution
 from aw.model.job_credential import BaseJobCredentials, JobUserCredentials, JobGlobalCredentials
 from aw.model.permission import CHOICE_PERMISSION_READ, CHOICE_PERMISSION_WRITE, CHOICE_PERMISSION_DELETE
 from aw.api_endpoints.base import API_PERMISSION, get_api_user, GenericResponse, BaseResponse, api_docs_delete, \
-    api_docs_put, api_docs_post
+    api_docs_put, api_docs_post, validate_no_xss
 from aw.utils.permission import has_credentials_permission, has_manager_privileges
 from aw.config.hardcoded import SECRET_HIDDEN
 from aw.utils.util import is_null
@@ -50,11 +50,25 @@ class JobGlobalCredentialsWriteRequest(serializers.ModelSerializer):
     connect_pass = serializers.CharField(max_length=100, required=False, default=None, allow_blank=True)
     ssh_key = serializers.CharField(max_length=5000, required=False, default=None, allow_blank=True)
 
+    def validate(self, attrs: dict):
+        for field in JobGlobalCredentials.api_fields_write:
+            if field in attrs and field not in BaseJobCredentials.SECRET_ATTRS:
+                validate_no_xss(value=attrs[field], field=field)
+
+        return attrs
+
 
 class JobUserCredentialsWriteRequest(JobGlobalCredentialsWriteRequest):
     class Meta:
         model = JobUserCredentials
         fields = JobUserCredentials.api_fields_write
+
+    def validate(self, attrs: dict):
+        for field in JobUserCredentials.api_fields_write:
+            if field in attrs and field not in BaseJobCredentials.SECRET_ATTRS:
+                validate_no_xss(value=attrs[field], field=field)
+
+        return attrs
 
 
 def are_global_credentials(request) -> bool: