+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 0.0.21
 
-* ...
+* Added [validation against XSS](https://github.com/ansibleguy/webui/issues/44)
 
 ----